Refact of WrapperHashtagHistory
