+++ src/components/HashtagGame/index.js
@@ -1,59 +1,21 @@
-import React, { useState } from "react";
+import React from "react";
 import "./styles.css";
 import BoardGame from "../../objects/BoardGame";
 import PlayerGame from "../../objects/PlayerGame";
 
-const HashtagGame = ({ callback }) => {
-    const [nextPlayer, setNextPlayer] = useState("x");
-
-    const [lastRound, setLastRound] = useState(0);
-
-    const [history, setHistory] = useState([
-        {
-            round: 0,
-            state: [
-                { id: 1, content: "" },
-                { id: 2, content: "" },
-                { id: 3, content: "" },
-                { id: 4, content: "" },
-                { id: 5, content: "" },
-                { id: 6, content: "" },
-                { id: 7, content: "" },
-                { id: 8, content: "" },
-                { id: 9, content: "" }
-            ]
-        }
-    ]);
-
-    const handleClick = id => {
-        setHistory(old => [
-            ...old,
-            {
-                round: lastRound + 1,
-                state: old[lastRound].state.map(player =>
-                    player.id === id ? { id, content: nextPlayer } : player
-                )
-            }
-        ]);
-        setLastRound(old => old + 1);
-        callback(nextPlayer);
-        setNextPlayer(old => (old === "x" ? "o" : "x"));
-    };
-
-    return (
-        <BoardGame>
-            <ul className="hashtag-game">
-                {history[lastRound].state.map(({ id, content }) => (
-                    <li
-                        key={id}
-                        className="item"
-                        onClick={() => content === "" && handleClick(id)}>
-                        <PlayerGame id={id} content={content} />
-                    </li>
-                ))}
-            </ul>
-        </BoardGame>
-    );
-};
+const HashtagGame = ({ historyGame, onClick, lastRound }) => (
+    <BoardGame>
+        <ul className="hashtag-game">
+            {historyGame[lastRound].state.map(({ id, content }) => (
+                <li
+                    key={id}
+                    className="item"
+                    onClick={() => content === "" && onClick(id)}>
+                    <PlayerGame id={id} content={content} />
+                </li>
+            ))}
+        </ul>
+    </BoardGame>
+);
 
 export default HashtagGame;
